requirements.txt
kafka-python==2.0.2
requests==2.27.1
json5==0.9.6
numpy==1.24.3
snowflake-connector-python==3.5.0
snowflake-snowpark-python==1.8.0
pandas==1.3.3